Narrative Tool Update
Now checks to make sure player isn't already in a convo as well as not in frog form.
Now has an option for default message to show when entering the zone after already having spoken before.
Now added an option for timer to be used rather than user input to continue conversation.